packfile: use size_t for crc_offset
The crc_offset member of struct packed_git introduced by
629dffc461 ("packfile: compute and use the index CRC offset",
2020-05-25) stores the offset to the start of the crc table in a v2
index file as a 32 bit unsigned integer. However if the packfile
associated with the index contains more than 214748313 objects then
the offset to the crc table will be greater that 2^32 and so the
stored offset will be truncated. Fix this by using size_t instead and
move the definition up to avoid introducing any more padding in the
structure. As noted in f86f769550 ("compute pack .idx byte offsets
using size_t", 2020-11-13) on 32 bit systems size_t is still too small
to hold the correct offset but we will already have failed to mmap the
index file in that case.

As well as the structure member being too narrow to hold the correct
offset the offset calculation was truncated because nr and hashsz are
32 bit unsigned integers, fix this by casting nr to size_t.

+++ packfile.c
@@ -178,7 +178,7 @@ int load_idx(const char *path, const unsigned int hashsz, void *idx_map,
 		     */
 		    (sizeof(off_t) <= 4))
 			return error("pack too large for current definition of off_t in %s", path);
-		p->crc_offset = 8 + 4 * 256 + nr * hashsz;
+		p->crc_offset = 8 + 4 * 256 + (size_t)nr * hashsz;
 	}
 
 	p->index_version = version;
